+++ inst/tinytest/test_show.R
@@ -1,4 +1,4 @@
-Sys.setenv(LANGUAGE = "en") # Force locale
+Sys.setlocale("LC_MESSAGES", 'en_GB.UTF-8') # Force locale
 options(aion.calendar = calendar("CE"))
 
 # Format =======================================================================